requirements.txt
woocommerce
python-dateutil